Rename package spinner to ja-spinner
